Safari-609.1.11.1/LayoutTests/platform/ios/webaudio/audiobuffersource-exception-expected.txt

Tests that AudioBufferSourceNode validates start and stop calls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS bufferSource.stop(0) threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS bufferSource.start(0) threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS bufferSource.stop(0) threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS successfullyParsed is true

TEST COMPLETE
